Deleted testing statements, added comments
Got rid of various printf statements used for debugging. Also got rid of
the no-longer-used #if/#endif that made it not download the specified
file.

Then I added some printf statements to output roughly what's going on in
the program. Things should basically fly by, still, but it should be
nice in case it gets hung up for some reason.